@@ -19,11 +19,11 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: Implementation :: CPython",
 ]
 packages = [
@@ -32,31 +32,35 @@ packages = [
 
 
 [tool.poetry.dependencies]
-python = "<3.12,>=3.8.1"
-singer-sdk = "~=0.31.0"
-psycopg2-binary = "2.9.7"
+python = "<3.13,>=3.8.1"
+psycopg2-binary = "2.9.9"
 sqlalchemy = "<2"
 sshtunnel = "0.4.0"
 
-[tool.poetry.group.dev.dependencies]
-black = "^23.1.0"
-faker = ">=18.5.1,<20.0.0"
-flake8 = "^6.0.0"
-isort = "^5.10.1"
-mypy = "1.5.1"
-pre-commit = "^3.0.4"
-pydocstyle = "^6.1.1"
-singer-sdk = {version = "*", extras = ["testing"]}
-tox = "^4"
+[tool.poetry.dependencies.singer-sdk]
+version = "~=0.35.0"
 
-[tool.isort]
-profile = "black"
-multi_line_output = 3 # Vertical Hanging Indent
-src_paths = "tap_postgres"
+[tool.poetry.group.dev.dependencies]
+faker = ">=18.5.1"
+mypy = ">=1.8.0"
+pendulum = "~=3.0"
+pre-commit = ">=3.0.4"
+ruff = "~=0.2.0"
+pydocstyle = ">=6.1.1"
+tox = "~=4"
+types-paramiko = ">=3.3.0.0"
+types-simplejson = ">=3.19.0.2"
+types-sqlalchemy = ">=1.4.53.38"
+types-jsonschema = ">=4.19.0.3"
+types-psycopg2 = ">=2.9.21.20240118"
+
+[tool.poetry.dev-dependencies.singer-sdk]
+version = "*"
+extras = ["testing"]
 
 [tool.mypy]
 exclude = "tests"
-python_version = "3.9"
+python_version = "3.11"
 warn_unused_configs = true
 warn_unused_ignores = true
 
@@ -67,9 +71,51 @@ module = [
 ]
 
 [build-system]
-requires = ["poetry-core>=1.0.8", "poetry-dynamic-versioning"]
+requires = ["poetry-core==1.8.1", "poetry-dynamic-versioning==1.2.0"]
 build-backend = "poetry_dynamic_versioning.backend"
 
 [tool.poetry.scripts]
 # CLI declaration
 tap-postgres = 'tap_postgres.tap:TapPostgres.cli'
+
+[tool.poetry-dynamic-versioning]
+enable = true
+vcs = "git"
+style = "pep440"
+
+[tool.ruff]
+target-version = "py38"
+
+[tool.ruff.lint]
+select = [
+  "F",    # Pyflakes
+  "W",    # pycodestyle warnings
+  "E",    # pycodestyle errors
+  "I",    # isort
+  "N",    # pep8-naming
+  "D",    # pydocsyle
+  "UP",   # pyupgrade
+  "ICN",  # flake8-import-conventions
+  "RET",  # flake8-return
+  "SIM",  # flake8-simplify
+  "TCH",  # flake8-type-checking
+  "ERA",  # eradicate
+  "PGH",  # pygrep-hooks
+  "PL",   # Pylint
+  "PERF", # Perflint
+  "RUF",  # ruff
+]
+
+[tool.ruff.lint.flake8-import-conventions]
+banned-from = ["sqlalchemy"]
+
+[tool.ruff.lint.flake8-import-conventions.extend-aliases]
+sqlalchemy = "sa"
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*" = [
+  "D",
+]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
